Move _expand_dims to funcs.py module
